test(TLS): check TLS12 connection
