+++ composer.json
@@ -24,7 +24,10 @@
       ],
       "Commercetools\\Core\\Fixtures\\": [
         "tests/fixtures/"
-      ]
+      ],
+      "Commercetools\\Core\\Test\\Functional": [
+        "tests/functional/"
+      ],
     }
   },
   "require": {
