+++ phpunit.xml.dist
@@ -11,6 +11,9 @@
         <testsuite name="integration">
             <directory suffix="Test.php">tests/integration</directory>
         </testsuite>
+        <testsuite name="functional">
+            <directory suffix="Test.php">tests/functional</directory>
+        </testsuite>
     </testsuites>
 
     <filter>
